[Done] Diffusion subscriptions commands

+++ isartbot/ext/diffusion.py
@@ -26,6 +26,7 @@ import discord
 
 from discord.ext import commands
 
+from isartbot.helper     import Helper
 from isartbot.checks     import is_super_admin, is_moderator
 from isartbot.database   import Diffusion, DiffusionOperator, DiffusionSubscription
 from isartbot.converters import DiffusionConverter
@@ -46,10 +47,16 @@ class DiffusionExt(commands.Cog):
         """ Diffusion command group """
         pass
 
-    @diffusion.group(pass_context=True, help="diffusion_operator_help", description="diffusion_operator_description")
+    @diffusion.group(aliases = ["op"], pass_context=True, help="diffusion_operator_help", description="diffusion_operator_description")
     @commands.check(is_super_admin)
     async def operator(self, ctx):
         """ Diffusion operator command group """
+        pass
+
+    @diffusion.group(aliases = ["sub"], pass_context=True, help="diffusion_subscription_help", description="diffusion_subscription_description")
+    @commands.check(is_moderator)
+    async def subscription(self, ctx):
+        """ Diffusion subscription command group """
         pass
 
     # Commands
@@ -69,13 +76,7 @@ class DiffusionExt(commands.Cog):
         self.bot.database.session.add(new_diffusion)
         self.bot.database.session.commit()
 
-        translations = await self.bot.get_translations(ctx, ["success_title", "diffusion_created"])
-        embed = discord.Embed(
-            title       = translations["success_title"],
-            description = translations["diffusion_created"].format(diffusion_name),
-            color       = discord.Color.green()
-        )
-        await ctx.send(embed=embed)
+        await Helper.send_success(ctx, ctx.channel, "diffusion_created", format_content = (diffusion_name))
 
     @diffusion.command(help="diffusion_delete_help", description="diffusion_delete_description")
     @commands.check(is_super_admin)
@@ -84,20 +85,19 @@ class DiffusionExt(commands.Cog):
 
         # Checking if the diffusion exists
         if (diffusion == None):
-            return await self.diffusion_does_not_exists_error(ctx, diffusion.name)
+            return await self.diffusion_does_not_exists_error(ctx)
+
+        # Asking for confirmation
+        result = await Helper.ask_confirmation(ctx, ctx.channel, "diffusion_delete_title",
+            initial_content = "diffusion_delete_content", initial_format = (diffusion.name),
+            success_content = "diffusion_deleted"       , success_format = (diffusion.name),
+            failure_content = "diffusion_delete_aborted")
+        if (not result):
+            return
 
         # Otherwise, deleting the diffusion
         self.bot.database.session.delete(diffusion)
         self.bot.database.session.commit()
-
-        translations = await self.bot.get_translations(ctx, ["success_title", "diffusion_deleted"])
-        embed = discord.Embed(
-            title       = translations["success_title"],
-            description = translations["diffusion_deleted"].format(diffusion.name),
-            color       = discord.Color.green()
-        )
-
-        await ctx.send(embed=embed)
 
     @diffusion.command(name="list", help="diffusion_list_help", description="diffusion_list_description")
     async def diffusion_list(self, ctx):
@@ -116,7 +116,7 @@ class DiffusionExt(commands.Cog):
         """ Adds a new diffusion operator to the selected diffusion """
         # Checking if the diffusion exists
         if (diffusion == None):
-            return await self.diffusion_does_not_exists_error(ctx, diffusion.name)
+            return await self.diffusion_does_not_exists_error(ctx)
 
         # Checking if this member is already an operator
         operator = self.bot.database.session.query(DiffusionOperator.discord_id).\
@@ -130,14 +130,7 @@ class DiffusionExt(commands.Cog):
         self.bot.database.session.add(DiffusionOperator(diffusion = diffusion, discord_id = new_operator.id))
         self.bot.database.session.commit()
 
-        translations = await self.bot.get_translations(ctx, ["success_title", "diffusion_operator_added"])
-        embed = discord.Embed(
-            title       = translations["success_title"],
-            description = translations["diffusion_operator_added"].format(new_operator.mention, diffusion.name),
-            color       = discord.Color.green()
-        )
-
-        await ctx.send(embed=embed)
+        await Helper.send_success(ctx, ctx.channel, "diffusion_operator_added", format_content = (new_operator.mention, diffusion.name))
 
     @operator.command(help="diffusion_operator_remove_help", description="diffusion_operator_remove_description")
     async def remove(self, ctx, diffusion: DiffusionConverter, old_operator: discord.Member):
@@ -145,7 +138,7 @@ class DiffusionExt(commands.Cog):
 
         # Checking if the diffusion exists
         if (diffusion == None):
-            return await self.diffusion_does_not_exists_error(ctx, diffusion.name)
+            return await self.diffusion_does_not_exists_error(ctx)
 
         # Checking if this member is already an operator
         operator = self.bot.database.session.query(DiffusionOperator).\
@@ -155,18 +148,16 @@ class DiffusionExt(commands.Cog):
         if (operator == None):
             return await self.member_not_operator(ctx, old_operator, diffusion.name)
 
+        # Asking for confirmation
+        result = await Helper.ask_confirmation(ctx, ctx.channel, "diffusion_operator_remove_title",
+            initial_content = "diffusion_operator_remove_content", initial_format = (old_operator.mention, diffusion.name),
+            success_content = "diffusion_operator_removed"       , success_format = (old_operator.mention, diffusion.name),
+            failure_content = "diffusion_operator_remove_content_failure")
+        if (not result):
+            return
+
         self.bot.database.session.delete(operator)
         self.bot.database.session.commit()
-
-        translations = await self.bot.get_translations(ctx, ["success_title", "diffusion_operator_removed"])
-        embed = discord.Embed(
-            title       = translations["success_title"],
-            description = translations["diffusion_operator_removed"].format(old_operator.mention, diffusion.name),
-            color       = discord.Color.green()
-        )
-
-        await ctx.send(embed=embed)
-
 
     @operator.command(name="list", help="diffusion_operator_list_help", description="diffusion_operator_list_description")
     async def operator_list(self, ctx, diffusion: DiffusionConverter):
@@ -174,7 +165,7 @@ class DiffusionExt(commands.Cog):
 
         # Checking if the diffusion exists
         if (diffusion == None):
-            return await self.diffusion_does_not_exists_error(ctx, diffusion.name)
+            return await self.diffusion_does_not_exists_error(ctx)
 
         operators_id = self.bot.database.session.query(DiffusionOperator.discord_id).\
             filter(DiffusionOperator.diffusion == diffusion).all()
@@ -187,75 +178,116 @@ class DiffusionExt(commands.Cog):
 
         await ctx.send(embed=embed)
 
-
     @diffusion.command(help="diffusion_diffuse_help", description="diffusion_diffuse_description")
-    async def diffuse(self, ctx, diffusion: DiffusionConverter, message: discord.Message):
+    async def diffuse(self, ctx, diffusion: DiffusionConverter, *, message: str):
         """ Diffuses a messages to all subscribers """
         pass
 
-    @diffusion.command(help="diffusion_subscribe_help", description="diffusion_subscribe_description")
-    @commands.check(is_moderator)
+    @subscription.command(aliases = ["sub", "add", "create"], help="diffusion_subscription_subscribe_help", 
+        description="diffusion_subscription_subscribe_description")
     async def subscribe(self, ctx, diffusion: DiffusionConverter, channel: discord.TextChannel, diffusion_tag: discord.Role = None):
         """ Subscribes to a diffusion """
-        pass
+    
+        # Checking if the diffusion exists
+        if (diffusion == None):
+            return await self.diffusion_does_not_exists_error(ctx)
 
-    @diffusion.command(help="diffusion_unsubscribe_help", description="diffusion_unsubscribe_description")
-    @commands.check(is_moderator)
-    async def unsubscribe(self, ctx, diffusion: DiffusionConverter):
+        # Checking if the subscription for this channel already exists
+        count = self.bot.database.session.query(DiffusionSubscription).\
+            filter(DiffusionSubscription.diffusion          == diffusion,
+                   DiffusionSubscription.discord_channel_id == channel.id).\
+            count()
+        
+        if (count > 0):
+            return await self.subscription_already_exists(ctx)
+
+        # Checking if we have enough permissions in order to operate the starboard on the designated channel
+        required_perms = discord.Permissions().none()
+        required_perms.update(send_messages=True)
+        if (not required_perms.is_subset(channel.permissions_for(ctx.guild.me))):
+            raise commands.BotMissingPermissions(["send_messages"])
+
+        new_subscription = DiffusionSubscription(
+                discord_server_id  = ctx.guild.id,
+                discord_channel_id = channel.id,
+                tag                = diffusion_tag.mention if diffusion_tag != None else "",
+                diffusion          = diffusion
+            )
+
+        self.bot.database.session.add(new_subscription)
+        self.bot.database.session.commit()
+
+        await Helper.send_success(ctx, ctx.channel, "diffusion_subscribe_success")
+
+    @subscription.command(aliases = ["unsub", "remove", "delete"], help="diffusion_subscription_unsubscribe_help",
+        description="diffusion_subscription_unsubscribe_description")
+    async def unsubscribe(self, ctx, diffusion: DiffusionConverter, channel: discord.TextChannel):
         """ Unsubscribes off a diffusion """
-        pass
+        
+        # Checking if the diffusion exists
+        if (diffusion == None):
+            return await self.diffusion_does_not_exists_error(ctx)
+
+        subscription = self.bot.database.session.query(DiffusionSubscription).\
+            filter(DiffusionSubscription.diffusion          == diffusion,
+                   DiffusionSubscription.discord_channel_id == channel.id).\
+            first()
+
+        if (subscription == None):
+            return await self.no_such_subscription_error(ctx)
+
+        self.bot.database.session.delete(subscription)
+        self.bot.database.session.commit()
+
+        await Helper.send_success(ctx, ctx.channel, "diffusion_unsubscribe_success")
+
+    @subscription.command(name = "list", help="diffusion_subscription_list_help", description="diffusion_subscription_list_description")
+    async def subscription_list(self, ctx, channel: discord.TextChannel = None):
+        """ Lists all the subscriptions in a channel or server """
+        
+        subscriptions = self.bot.database.session.query(DiffusionSubscription).\
+            filter(DiffusionSubscription.discord_server_id == ctx.guild.id).\
+            all()
+
+        if len(subscriptions) > 0:
+            descritpiton = '\n'.join([f"\u2022 {sub.diffusion.name} -> {(await self.bot.fetch_channel(sub.discord_channel_id)).mention}" for sub in subscriptions])
+        else:
+            descritpiton = await ctx.bot.get_translation(ctx, 'diffusion_subscription_list_empty')
+
+        embed = discord.Embed()
+
+        embed.title       = await ctx.bot.get_translation(ctx, 'diffusion_subscription_list_title')
+        embed.description = descritpiton
+        embed.colour      = discord.Color.green()
+
+        await ctx.send(embed=embed)
 
     # Error handlers
 
-    async def diffusion_does_not_exists_error(self, ctx, diffusion_name: str):
+    async def diffusion_does_not_exists_error(self, ctx):
         """ Diffusion does not exists error """
-
-        translations = await self.bot.get_translations(ctx, ["failure_title", "diffusion_doesnt_exists"])
-        embed = discord.Embed(
-            title       = translations["failure_title"],
-            description = translations["diffusion_doesnt_exists"].format(diffusion_name, ctx.prefix),
-            color       = discord.Color.red()
-        )
-
-        await ctx.send(embed=embed)
-
+        await Helper.send_error(ctx, ctx.channel, "diffusion_doesnt_exists", format_content = (ctx.prefix))
 
     async def diffusion_already_exists_error(self, ctx, diffusion_name: str):
         """ Diffusion already exists error """
-
-        translations = await self.bot.get_translations(ctx, ["failure_title", "diffusion_already_exists"])
-        embed = discord.Embed(
-            title       = translations["failure_title"],
-            description = translations["diffusion_already_exists"].format(diffusion_name),
-            color       = discord.Color.red()
-        )
-
-        await ctx.send(embed=embed)
-
+        await Helper.send_error(ctx, ctx.channel, "diffusion_already_exists", format_content = (diffusion_name))
 
     async def member_is_already_operator(self, ctx, operator: discord.Member, diffusion_name: str):
         """ Member is already an operator error """
-
-        translations = await self.bot.get_translations(ctx, ["failure_title", "diffusion_already_operator"])
-        embed = discord.Embed(
-            title       = translations["failure_title"],
-            description = translations["diffusion_already_operator"].format(operator.mention, diffusion_name),
-            color       = discord.Color.red()
-        )
-
-        await ctx.send(embed=embed)
+        await Helper.send_error(ctx, ctx.channel, "diffusion_already_operator", format_content = (operator.mention, diffusion_name))
 
     async def member_not_operator(self, ctx, member: discord.Member, diffusion_name: str):
         """ Member isn't an operator error """
+        await Helper.send_error(ctx, ctx.channel, "diffusion_not_operator", format_content = (member.mention, diffusion_name))
 
-        translations = await self.bot.get_translations(ctx, ["failure_title", "diffusion_not_operator"])
-        embed = discord.Embed(
-            title       = translations["failure_title"],
-            description = translations["diffusion_not_operator"].format(member.mention, diffusion_name),
-            color       = discord.Color.red()
-        )
+    async def subscription_already_exists(self, ctx):
+        """ Subscription already exists error """
+        await Helper.send_error(ctx, ctx.channel, "diffusion_subscription_already_exists", format_content = (ctx.prefix))
 
-        await ctx.send(embed=embed)
+    async def no_such_subscription_error(self, ctx):
+        """ No such subscription error """
+        await Helper.send_error(ctx, ctx.channel, "diffusion_no_such_subscription", format_content = (ctx.prefix))
+
 
 def setup(bot):
     bot.add_cog(DiffusionExt(bot))
